Split hosted resource catalog from workspace
